feat(executor): add plan info for processor profile
